backend/requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.53.0
openai==1.93.2
serpapi==0.1.5
python-dotenv==1.0.0
aiofiles==23.2.1
spacy==3.7.2
nltk==3.8.1 